rename AngeGraphModel to AngelGraphModel
