+++ core/src/main/scala/com/tencent/angel/sona/core/GraphModelPool.scala
@@ -2,22 +2,22 @@ package com.tencent.angel.sona.core
 
 import com.tencent.angel.mlcore.conf.SharedConf
 import com.tencent.angel.mlcore.variable.VarState
-import com.tencent.angel.sona.ml.AngeGraphModel
+import com.tencent.angel.sona.ml.AngelGraphModel
 import org.apache.spark.TaskContext
 
 import scala.collection.mutable
 
 class GraphModelPool(sparkEnvContext: SparkEnvContext, numTask: Int) {
-  @transient private lazy val modelQueue = new mutable.Queue[AngeGraphModel]()
-  @transient private lazy val usedMap = new mutable.HashMap[Long, AngeGraphModel]()
+  @transient private lazy val modelQueue = new mutable.Queue[AngelGraphModel]()
+  @transient private lazy val usedMap = new mutable.HashMap[Long, AngelGraphModel]()
 
-  def borrowModel(conf: SharedConf): AngeGraphModel = this.synchronized {
+  def borrowModel(conf: SharedConf): AngelGraphModel = this.synchronized {
     val partitionId = TaskContext.getPartitionId()
 
     if (usedMap.contains(partitionId)) {
       usedMap(partitionId)
     } else if (modelQueue.isEmpty) {
-      val model = new AngeGraphModel(conf, numTask)
+      val model = new AngelGraphModel(conf, numTask)
       model.buildNetwork()
       model.createMatrices(sparkEnvContext)
       model.setState(VarState.Initialized)
@@ -32,7 +32,7 @@ class GraphModelPool(sparkEnvContext: SparkEnvContext, numTask: Int) {
     }
   }
 
-  def returnModel(model: AngeGraphModel): Unit = this.synchronized {
+  def returnModel(model: AngelGraphModel): Unit = this.synchronized {
     val partitionId = TaskContext.getPartitionId()
     require(usedMap.contains(partitionId), "Error!")
     modelQueue.enqueue(model)
